+++ sdk/resourcemanager/providerhub/armproviderhub/ze_generated_example_customrollouts_client_test.go
@@ -78,7 +78,7 @@ func ExampleCustomRolloutsClient_CreateOrUpdate() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/providerhub/resource-manager/Microsoft.ProviderHub/stable/2020-11-20/examples/CustomRollouts_ListByProviderRegistration.json
-func ExampleCustomRolloutsClient_ListByProviderRegistration() {
+func ExampleCustomRolloutsClient_NewListByProviderRegistrationPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -90,7 +90,7 @@ func ExampleCustomRolloutsClient_ListByProviderRegistration() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByProviderRegistration("<provider-namespace>",
+	pager := client.NewListByProviderRegistrationPager("<provider-namespace>",
 		nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)