+++ sdk/resourcemanager/providerhub/armproviderhub/ze_generated_example_notificationregistrations_client_test.go
@@ -112,7 +112,7 @@ func ExampleNotificationRegistrationsClient_Delete() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/providerhub/resource-manager/Microsoft.ProviderHub/stable/2020-11-20/examples/NotificationRegistrations_ListByProviderRegistration.json
-func ExampleNotificationRegistrationsClient_ListByProviderRegistration() {
+func ExampleNotificationRegistrationsClient_NewListByProviderRegistrationPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -124,7 +124,7 @@ func ExampleNotificationRegistrationsClient_ListByProviderRegistration() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByProviderRegistration("<provider-namespace>",
+	pager := client.NewListByProviderRegistrationPager("<provider-namespace>",
 		nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)