+++ sdk/resourcemanager/providerhub/armproviderhub/zz_generated_operations_client.go
@@ -147,10 +147,10 @@ func (client *OperationsClient) deleteCreateRequest(ctx context.Context, provide
 	return req, nil
 }
 
-// List - Lists all the operations supported by Microsoft.ProviderHub.
+// NewListPager - Lists all the operations supported by Microsoft.ProviderHub.
 // If the operation fails it returns an *azcore.ResponseError type.
 // options - OperationsClientListOptions contains the optional parameters for the OperationsClient.List method.
-func (client *OperationsClient) List(options *OperationsClientListOptions) *runtime.Pager[OperationsClientListResponse] {
+func (client *OperationsClient) NewListPager(options *OperationsClientListOptions) *runtime.Pager[OperationsClientListResponse] {
 	return runtime.NewPager(runtime.PageProcessor[OperationsClientListResponse]{
 		More: func(page OperationsClientListResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0